refactor: axiosInstance 및 타입 파일로 별도 분리 (#10)
