+++ src/api/apiRequest.ts
@@ -1,23 +1,8 @@
-import axios, { AxiosInstance } from 'axios';
-import { getAccessToken, setAccessToken } from './auth/token';
+import axios from 'axios';
+import { getAccessToken, setAccessToken } from '../utils/tokenHandler';
 import { reissueToken } from './auth/reissue';
-
-interface RequestConfig {
-    requireToken?: boolean;
-    url: string;
-    method?: string; // default GET으로 할거라 일단 optional
-    data?: unknown;
-    headers?: Record<string, string>;
-}
-
-export const axiosInstance: AxiosInstance = axios.create({
-    baseURL: import.meta.env.VITE_BASE_URL,
-    withCredentials: false, // jwt 쿠키로 관리 안해서 false로
-    headers: {
-        'Content-Type': 'application/json',
-    },
-    timeout: 3000,
-});
+import { axiosInstance } from './axiosInstance';
+import { RequestConfig } from '@/types/api.types';
 
 /**
  * API 호출할 때 최종적으로 사용할 함수
@@ -37,6 +22,7 @@ export const apiRequest = async ({
     data,
     headers = {},
     requireToken = false,
+    clubName,
 }: RequestConfig) => {
     try {
         if (requireToken) {
@@ -63,22 +49,22 @@ export const apiRequest = async ({
             console.log(error.response?.data.message, '라는 이유로 다시 요청!');
             // 401일 때(토큰이 만료됐을 때)는 토큰 재발급하고 이전에 했던 작업 그대로 하도록
             // axios에러면서 응답이 401일 때만 토큰 관련 헨들링 진행
-            try {
-                const newToken = await reissueToken(); // 나중에 개발 완료되면 바뀐다고 하네여
-                if (!newToken) throw new Error('토큰 재발급 실패');
+            // try {
+            //     const newToken = await reissueToken(clubName); // 나중에 개발 완료되면 바뀐다고 하네여
+            //     if (!newToken) throw new Error('토큰 재발급 실패');
 
-                return axiosInstance({
-                    url,
-                    method,
-                    data,
-                    headers: {
-                        ...headers,
-                        Authorization: `Bearer ${newToken}`,
-                    },
-                });
-            } catch (error) {
-                console.error(error);
-            }
+            //     return axiosInstance({
+            //         url,
+            //         method,
+            //         data,
+            //         headers: {
+            //             ...headers,
+            //             Authorization: `Bearer ${newToken}`,
+            //         },
+            //     });
+            // } catch (error) {
+            //     console.error(error);
+            // }
         }
         // axios에러가 아닌 경우 (ex. 문법 오류, 타입 오류 등등..)로 구분
         console.error(error);
